Two changes, the first change in lbp is for fixing an error marked when the lbp operator try to change the value from int64 to uint32, in newer version of numpy this action is forbidden, the second change is a modification in SVM classifier to save and load the classifier. The modification works with default params, I think this can be useful

+++ py/facerec/lbp.py
@@ -94,7 +94,7 @@ class ExtendedLBP(LocalDescriptor):
             N += w4*X[cy:cy+dy,cx:cx+dx]
             # update LBP codes        
             D = N >= C
-            result += (1<<i)*D
+            result += ((1<<i)*D).astype("uint32")
         return result
 
     @property
